Extract BuildAndStartProducerSideServer
We extracted the `BuildAndStartProducerSideServer` method from the
Service module and moved it in to the ProducerSideService module.
This makes the CloudCollector able to reuse the
`BuildAndStartProducerSideServer` method.

Bug: http://b/218500565
